first cut at weaving coverage into csv.out 	modified:   ../src/array_api.py 	modified:   ../src/lib.py 	modified:   ../src/redux.py
